requirements.txt
apprise==1.3.0
daemonize==2.5.0
fire==0.5.0
Mastodon.py==1.8.0
python-dotenv==1.0.0
